Improved support of non-UFO3 defcon glyph objects
MutatorMath really needs to run with the UFO3 branch of defcon. For
experimentation and development however it might be useful to have some
level of compatibility with the UFO2 defcon objects. So this wraps the
main differences (round() method *1 and onlyGeometry argument in
extractGlyph() method *2). The results will be better using UFO3*, but
at least your test scripts won’t crash.

*1 All geometry data will not be rounded.
*2 The onlyGeometry argument prevents the instance glyph from picking
up odd unicode values from a master (in some cases)