ufo.document: don't use logging.basicConfig; only configure the 'mutatorMath' logger
the basicConfig function configures the "root" logger, i.e. the parent of all loggers, so it also influences loggers from other libraries which are outside of the 'mutatorMath' logging namespace.

If we want to configure a default file handler for mutatorMath, then we must do it for the specific 'mutatorMath' logger only, and not for the global, root logger.

E.g., see: https://github.com/googlei18n/fontmake/pull/66